[Chore #160364279] Restructure js source
