+++ src/app/components/auth/SignIn.js
@@ -1,9 +1,9 @@
-import API from "../API";
-import DOM from "../DOM";
-import Auth from "./Auth";
-import Error from "../Error";
-import Router from "../Router";
-import Message from "../Message";
+import API from "../../modules/API";
+import DOM from "../../modules/DOM";
+import Auth from "../../modules/Auth";
+import Error from "../../modules/Error";
+import Router from "../../modules/Router";
+import Message from "../../modules/Message";
 
 export default class SignIn {
     static init() {
